make theme work properly on sign in page
